+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [undecided] - Unreleased
+
+### Added
+- New `parentTheme` and `parentVersions` fields in the `ThemeDataType`
+
 ## [1.2.0] - 2024-11-27
 This is the stable release of v1.2.0. No changes have been made since v1.2.0-rc.1.